requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
httpx>=0.27
beautifulsoup4>=4.12
pdfplumber>=0.11
pypdf>=5.0
pymupdf>=1.24
openai>=1.55
pydantic>=2.8
sse-starlette>=2.1
rapidfuzz>=3.9
